+++ lib/kitchen/driver/rackspace.rb
@@ -36,14 +36,9 @@ module Kitchen
       default_config :wait_for, 600
       default_config :no_ssh_tcp_check, false
       default_config :no_ssh_tcp_check_sleep, 120
-
-      default_config :image_id do |driver|
-        driver.default_image
-      end
-
-      default_config :server_name do |driver|
-        driver.default_name
-      end
+      default_config(:image_id) { |driver| driver.default_image }
+      default_config(:server_name) { |driver| driver.default_name }
+      default_config :networks, nil
 
       default_config :public_key_path do
         [
@@ -107,33 +102,26 @@ module Kitchen
       private
 
       def compute
-        server_def = {
-          provider:           'Rackspace',
-          version:            config[:version],
-          rackspace_username: config[:rackspace_username],
-          rackspace_api_key:  config[:rackspace_api_key]
-        }
-        if config[:rackspace_region]
-          server_def[:rackspace_region] = config[:rackspace_region]
+        server_def = { provider: 'Rackspace' }
+        opts = [:version, :rackspace_username, :rackspace_api_key,
+                :rackspace_region]
+        opts.each do |opt|
+          server_def[opt] = config[opt]
         end
         Fog::Compute.new(server_def)
       end
 
       def create_server
-        compute.servers.bootstrap(
-          name:            config[:server_name],
-          image_id:        config[:image_id],
-          flavor_id:       config[:flavor_id],
-          public_key_path: config[:public_key_path]
-        )
+        server_def = { name: config[:server_name], networks: networks }
+        [:image_id, :flavor_id, :public_key_path].each do |opt|
+          server_def[opt] = config[opt]
+        end
+        compute.servers.bootstrap(server_def)
       end
 
       def images
         @images ||= begin
-          json_file = File.expand_path(
-            File.join(%w(.. .. .. .. data images.json)),
-            __FILE__
-          )
+          json_file = File.expand_path('../../../../data/images.json', __FILE__)
           JSON.load(IO.read(json_file))
         end
       end
@@ -141,10 +129,17 @@ module Kitchen
       def tcp_check(state)
         # allow driver config to bypass SSH tcp check -- because
         # it doesn't respect ssh_config values that might be required
-
         wait_for_sshd(state[:hostname]) unless config[:no_ssh_tcp_check]
         sleep(config[:no_ssh_tcp_check_sleep]) if config[:no_ssh_tcp_check]
         puts '(ssh ready)'
+      end
+
+      def networks
+        base_nets = %w(
+          00000000-0000-0000-0000-000000000000
+          11111111-1111-1111-1111-111111111111
+        )
+        config[:networks] ? base_nets + config[:networks] : nil
       end
     end
   end