+++ spec/kitchen/driver/rackspace_spec.rb
@@ -29,6 +29,7 @@ describe Kitchen::Driver::Rackspace do
   let(:config) { Hash.new }
   let(:state) { Hash.new }
   let(:platform_name) { 'ubuntu' }
+  let(:default_networks) { nil }
 
   let(:instance) do
     double(
@@ -102,6 +103,10 @@ describe Kitchen::Driver::Rackspace do
 
       it 'defaults the TCP bypass sleep time to 120 seconds' do
         expect(driver[:no_ssh_tcp_check_sleep]).to eq(120)
+      end
+
+      it 'defaults to the standard Rackspace networks' do
+        expect(driver[:networks]).to eq(default_networks)
       end
     end
 
@@ -212,6 +217,7 @@ describe Kitchen::Driver::Rackspace do
         driver.create(state)
       end
     end
+
   end
 
   describe '#destroy' do
@@ -313,7 +319,8 @@ describe Kitchen::Driver::Rackspace do
         server_name: 'hello',
         image_id: 'there',
         flavor_id: 'captain',
-        public_key_path: 'tarpals'
+        public_key_path: 'tarpals',
+        networks: default_networks
       }
     end
     before(:each) do
@@ -337,6 +344,40 @@ describe Kitchen::Driver::Rackspace do
 
     it 'creates the server using a compute connection' do
       expect(driver.send(:create_server)).to eq(@expected)
+    end
+
+    context 'additional networks specified' do
+      let(:server_id) { '12345' }
+      let(:server) do
+        double(id: 'test123', wait_for: true,
+               public_ip_address: '1.2.3.4')
+      end
+      let(:hostname) { 'example.com' }
+      let(:servers_double) { double('servers', bootstrap: server) }
+      let(:compute_double) { double(Fog::Compute, servers: servers_double) }
+      let(:state) { { server_id: server_id, hostname: hostname } }
+      let(:driver) do
+        d = Kitchen::Driver::Rackspace.new(config)
+        allow(d).to receive(:wait_for_sshd).with('1.2.3.4').and_return(true)
+        d.instance = instance
+        allow(Fog::Compute).to receive(:new).and_return(compute_double)
+        d
+      end
+      let(:user_specified_network) { 'bob_dole' }
+      let(:config) do
+        {
+          rackspace_username: 'monkey',
+          rackspace_api_key: 'potato',
+          rackspace_region: 'ord',
+          networks: [user_specified_network]
+        }
+      end
+      it 'has the user specified network, plus default Rackspace networks' do
+        driver.create(state)
+        expect(servers_double).to have_received(:bootstrap) do |arg|
+          expect(arg[:networks][2]).to eq user_specified_network
+        end
+      end
     end
   end
 
@@ -373,6 +414,27 @@ describe Kitchen::Driver::Rackspace do
         expect(driver).to_not receive(:wait_for_sshd)
         expect(driver).to receive(:sleep)
         driver.send(:tcp_check, state)
+      end
+    end
+  end
+
+  describe '#networks' do
+    context 'the default Rackspace networks' do
+      it 'returns nil so Fog will use the defaults' do
+        expect(driver.send(:networks)).to eq(nil)
+      end
+    end
+
+    context 'a custom Rackspace network' do
+      let(:config) { { networks: %w(abcdefg) } }
+
+      it 'returns the base networks plus the custom one' do
+        expected = %w(
+          00000000-0000-0000-0000-000000000000
+          11111111-1111-1111-1111-111111111111
+          abcdefg
+        )
+        expect(driver.send(:networks)).to eq(expected)
       end
     end
   end